Move component MetroCard from 'components/cards' to 'components/index/CardsReference'
