+++ pages/cards/_card.vue
@@ -41,7 +41,7 @@ import MonitoringConsultationDeskReportsNumberCard from '@/components/index/Card
 // 新型コロナコールセンター相談件数
 import TelephoneAdvisoryReportsNumberCard from '@/components/index/CardsReference/TelephoneAdvisoryReportsNumberCard.vue'
 // 都営地下鉄の利用者数の推移
-import MetroCard from '@/components/cards/MetroCard.vue'
+import MetroCard from '@/components/index/CardsReference/MetroCard.vue'
 // 都庁来庁者数の推移
 import AgencyCard from '@/components/cards/AgencyCard.vue'
 /* eslint-enable simple-import-sort/sort */
